try to solve overflow SUW

+++ npbenchmark-main/SDK.PCP/pcp_vector.h
@@ -53,7 +53,7 @@ private: // member variables.
     std::vector<int> uncovered_value;
     boost::dynamic_bitset<> uncovered_vertices;
 
-    int num_uncovered;
+    long long int num_uncovered;
     int best_num_uncovered;
     int prev_num_uncovered;
